Move metabase/lib/types to metabase-lib
