+++ frontend/src/metabase-lib/lib/queries/StructuredQuery.ts
@@ -28,11 +28,11 @@ import { AggregationOperator } from "metabase-types/types/Metadata";
 import { DatabaseEngine, DatabaseId } from "metabase-types/types/Database";
 import { TableId } from "metabase-types/types/Table";
 import { Column } from "metabase-types/types/Dataset";
-import { TYPE } from "metabase/lib/types";
 import {
   isVirtualCardId,
   getQuestionIdFromVirtualTableId,
 } from "metabase/lib/saved-questions";
+import { TYPE } from "metabase-lib/lib/types/utils/isa";
 import { fieldRefForColumn } from "metabase-lib/lib/queries/utils/dataset";
 import { isSegment } from "metabase-lib/lib/queries/utils/filter";
 import { getUniqueExpressionName } from "metabase-lib/lib/queries/utils/expression";
