+++ frontend/src/metabase-lib/lib/queries/structured/Aggregation.ts
@@ -1,11 +1,11 @@
 // eslint-disable-next-line @typescript-eslint/ban-ts-comment
 // @ts-nocheck
 import { t } from "ttag";
-import { TYPE } from "metabase/lib/types";
 import { Aggregation as AggregationObject } from "metabase-types/types/Query";
 import { AggregationOperator } from "metabase-types/types/Metadata";
 import { MetricId } from "metabase-types/types/Metric";
 import { FieldId } from "metabase-types/types/Field";
+import { TYPE } from "metabase-lib/lib/types/utils/isa";
 import * as AGGREGATION from "metabase-lib/lib/queries/utils/aggregation";
 import Filter from "metabase-lib/lib/queries/structured/Filter";
 import Metric from "metabase-lib/lib/metadata/Metric";
