+++ frontend/src/metabase-lib/lib/queries/utils/field-ref.js
@@ -1,6 +1,6 @@
 import _ from "underscore";
 
-import { TYPE } from "metabase/lib/types";
+import { TYPE } from "metabase-lib/lib/types/utils/isa";
 import Field from "metabase-lib/lib/metadata/Field";
 import { FieldDimension } from "metabase-lib/lib/Dimension";
 import * as Table from "./table";
