+++ frontend/src/metabase/admin/datamodel/containers/FieldApp.jsx
@@ -23,7 +23,6 @@ import AdminLayout from "metabase/components/AdminLayout";
 import { LeftNavPane, LeftNavPaneItem } from "metabase/components/LeftNavPane";
 
 import { is_coerceable, coercions_for_type } from "cljs/metabase.types";
-import { isFK } from "metabase/lib/types";
 
 import ColumnSettings from "metabase/visualizations/components/ColumnSettings";
 
@@ -40,6 +39,7 @@ import { isCurrency } from "metabase/lib/schema_metadata";
 import Databases from "metabase/entities/databases";
 import Tables from "metabase/entities/tables";
 import Fields from "metabase/entities/fields";
+import { isFK } from "metabase-lib/lib/types/utils/isa";
 import { rescanFieldValues, discardFieldValues } from "../field";
 import UpdateCachedFieldValues from "../components/UpdateCachedFieldValues";
 import FieldRemapping from "../components/FieldRemapping";
