+++ frontend/src/metabase/reference/components/FieldTypeDetail.jsx
@@ -5,11 +5,11 @@ import { getIn } from "icepick";
 import { t } from "ttag";
 import * as MetabaseCore from "metabase/lib/core";
 import { isNumericBaseType } from "metabase/lib/schema_metadata";
-import { isFK } from "metabase/lib/types";
 
 import Select from "metabase/core/components/Select";
 
 import D from "metabase/reference/components/Detail.css";
+import { isFK } from "metabase-lib/lib/types/utils/isa";
 
 const FieldTypeDetail = ({
   field,
