+++ frontend/src/metabase/visualizations/components/ObjectDetail/ObjectDetailsTable.tsx
@@ -8,8 +8,8 @@ import ExpandableString from "metabase/query_builder/components/ExpandableString
 import EmptyState from "metabase/components/EmptyState";
 
 import { isID } from "metabase/lib/schema_metadata";
-import { TYPE, isa } from "metabase/lib/types";
 import { formatValue, formatColumn } from "metabase/lib/formatting";
+import { TYPE, isa } from "metabase-lib/lib/types/utils/isa";
 
 import { OnVisualizationClickType } from "./types";
 import {
