+++ frontend/src/metabase/writeback/utils.ts
@@ -1,5 +1,3 @@
-import { TYPE } from "metabase/lib/types";
-
 import type {
   ActionDashboardCard,
   BaseDashboardOrderedCard,
@@ -7,6 +5,7 @@ import type {
   Database as IDatabase,
 } from "metabase-types/api";
 import type { SavedCard } from "metabase-types/types/Card";
+import { TYPE } from "metabase-lib/lib/types/utils/isa";
 import type Database from "metabase-lib/lib/metadata/Database";
 import type Field from "metabase-lib/lib/metadata/Field";
