+++ frontend/test/metabase/lib/data_grid.unit.spec.js
@@ -9,7 +9,7 @@ import {
   COLUMN_SHOW_TOTALS,
 } from "metabase/lib/data_grid";
 
-import { TYPE } from "metabase/lib/types";
+import { TYPE } from "metabase-lib/lib/types/utils/isa";
 
 const dimension = i => ({
   name: "D" + i,
