+++ frontend/test/metabase/visualizations/lib/timeseries.unit.spec.js
@@ -6,7 +6,7 @@ import {
 } from "metabase/visualizations/lib/timeseries";
 import { getVisualizationTransformed } from "metabase/visualizations";
 
-import { TYPE } from "metabase/lib/types";
+import { TYPE } from "metabase-lib/lib/types/utils/isa";
 import { StringColumn, NumberColumn } from "../__support__/visualizations";
 
 describe("visualization.lib.timeseries", () => {
